Fork react-aria-modal and port it to TypeScript. (#6)
It seems like there isn't any movement on react-aria-modal and it's hard to change because of its lack of tests, the fact that it's in JS (not TS), and the fact that it has dependencies strewn across multiple github repositories (some of which have bugs that need fixing, e.g. JustFixNYC/tenants2#298 (comment)).

This is basically an attempt to fork react-aria-modal and all its dependencies and convert them to TypeScript. I removed code that was specific to anything less than React 16.