Add CommonJS versions of some packages. (#22)
This fixes #21 by adding CommonJS versions of the `util` and `geosearch-requester` packages, under the `commonjs` directory of each package.  We're only adding CommonJS versions of these for now because they can be used from non-browser environments where it's more likely for a transpiler/bundler to be absent.

This was made using strategies outlined in https://2ality.com/2019/10/hybrid-npm-packages.html.