add storageFactory function to util packge (#41)
We have been getting errors when localStorage isn't available (browser settings). I found this helpful blog that discusses some of those common issues and shares a good solution of using a storageFactory that wraps all the operations of local/session storage, but also handles the cases where localStorage isn't supported.

This PR adds this function to the util package, so that we can use this across our tools. Right now org-site uses localStorage to keep the cookies-consent banner hidden once closed (JustFixNYC/justfix-website#234), and we'll need to do the same thing with sessionStorage for the wow big portfolios warning (JustFixNYC/who-owns-what#689).